Move bin obj dataset to its own file and fix some broken imports
